modified:   ../ProductOfFour.py Around line 34: clean up comments concerning file run location (linux vs. Windows)
Change the windows drive letter from hard-coded to variable 'drive'

Change debugging constants. Max interval changed to 200 (from 51), max
increment changed to 50 (from 21). This was to test multiline output in
the squares.txt file.

Around line 85: the setting of header1 for both bash and windows was
identical. Removed the if logic (commented out) and replace it with a
simple variable assignment.

Around line 108: clean up comment.

Remove commented code that was used to manually write CSV data. This had
been replaced by implementing the csv module. As the csv module code
was tested and works, this comment is superfluous.

Around line 304: A print statement was erroneously merged into a
comment. Fixed.

Around line 310: The 'with open' specified an incorrect directory
"output2", changed to "output"

All subsequent notes refer to the Analysis of Squares section of the
code.

For generated sheets, change how the headers are determined
Change name of variable "count" to outputLineCount,as the latter is more
descriptive.

generate sheet:
	Old code was in error when looping over the osq (O-rdered SQ-uares)
	dictionary. When running "for k,v in osq.items():", modifications had
	been made to the copy within the dictionary. Instead, the dictionary
	entries (list[]'s) are now copied to a separate variable (vCSV) which is then
	modified. This leaves the dictionary intact.

generate text (report):
	Copied osq dictionary entries to a separate list vTXT.
	The vTXT list is then sliced to v2, starting at the first
	starting-number/increment pair.
	v2 is extended to ensure that there are at least a multiple of 8
	entries. This if v2 has 10 entries, it is extended to 16. The
	extended entries all contain '-'.

	v2 is enumerated. Starting at every 8th entry (0, 8, 16, ...)
	the starting-number/increment pairs are processed for output
	four-at-a-time.