directory changes, maximum output file change
README.md: output directory name change from Output to Pof4_Output

ProductOfFour.py: removed hard coded directories and logic to indicate running in bash shell or windows terminal. Instead used py's 'os' module to determine the program's directory and create a Pof4_Output directory and write to that.
Removed an "ALL" for MAX_INCREMENT_FILES and instead check for non-numeric input to indcate  that the INCREMENT_MAX is to be used.